single file component: move Award.ts to Award.vue
