+++ package.json
@@ -58,8 +58,8 @@
     "precompile": "npm run makeJson",
     "cover": "nyc mocha --file build/tests/utils/setup.js --recursive build/tests",
     "precover": "tsc --build tests/tsconfig.json",
-    "fix": "eslint --fix src tests --ext ts",
-    "lint": "eslint --cache src tests --ext ts",
+    "fix": "eslint --fix src tests --ext ts --ext vue",
+    "lint": "eslint --cache src tests --ext ts --ext vue",
     "makeCss": "lessc src/styles/common.less build/styles.css",
     "makeJson": "node make_static_json.js",
     "start": "node build/src/server.js",
