+++ src/components/PlayerHome.ts
@@ -3,7 +3,7 @@ import Vue from 'vue';
 import {Board} from './Board';
 import {Card} from './card/Card';
 import {Milestone} from './Milestone';
-import {Award} from './Award';
+import Award from './Award.vue';
 import {PlayersOverview} from './overview/PlayersOverview';
 import {WaitingFor} from './WaitingFor';
 import {Sidebar} from './Sidebar';
@@ -68,7 +68,7 @@ export const PlayerHome = Vue.component('player-home', {
     'players-overview': PlayersOverview,
     'waiting-for': WaitingFor,
     'milestone': Milestone,
-    'award': Award,
+    Award,
     'sidebar': Sidebar,
     'colony': Colony,
     'log-panel': LogPanel,
@@ -242,7 +242,7 @@ export const PlayerHome = Vue.component('player-home', {
 
                     <div v-if="player.players.length > 1" class="player_home_block--milestones-and-awards">
                         <milestone :milestones_list="player.game.milestones" />
-                        <award :awards_list="player.game.awards" />
+                        <Award :awards_list="player.game.awards" />
                     </div>
                 </div>
 
@@ -359,7 +359,7 @@ export const PlayerHome = Vue.component('player-home', {
 
                 <div class="player_home_block" v-if="player.players.length > 1">
                     <milestone :show_scores="false" :milestones_list="player.game.milestones" />
-                    <award :show_scores="false" :awards_list="player.game.awards" />
+                    <Award :show_scores="false" :awards_list="player.game.awards" />
                 </div>
 
                 <div class="player_home_block player_home_block--turnorder nofloat" v-if="player.players.length>1">
